+++ app/src/main/java/com/example/demo/application/TaskCreator.java
@@ -1,11 +1,11 @@
 package com.example.demo.application;
 
 import com.example.demo.infrastructure.Task;
-import com.example.demo.infrastructure.TaskRepository;
+import com.example.demo.infrastructure.TaskRepositoryImp;
 
 public class TaskCreator {
 
-    TaskRepository taskRepository = new TaskRepository();
+    TaskRepositoryImp taskRepository = new TaskRepositoryImp();
 
     public Task createTask(String content) {
         return taskRepository.save(content);
